backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.0
alembic==1.11.1
python-dotenv==1.0.0
redis==4.5.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
pydantic==1.10.7
